Tidied up tests

+++ src/Main.hs
@@ -15,7 +15,7 @@ main  = do
                 if length a > 0
                         then return $ read (head a)
                         else return 10000      
-        putStrLn $ "Testing the word-count MapReduce algorithm with size bound " ++ show b
+        putStrLn $ "Testing the word-count MapReduce algorithm with size bound " ++ show (fromIntegral b)
         putStrLn "WARNING: Can be very slow if bound > 10000"          
         putStrLn "Total count test (coarse-grained): " >> quickCheck (prop_Equal b)
         putStrLn "Detail test (very strong): " >> quickCheck (prop_FullCheck b)
